chore: update structure
