+++ packages/vincent-contracts/src.ts/networks/vDatil/datil-mainnet/vincentContext.ts
@@ -1,16 +1,15 @@
 import { createWalletClient, http } from 'viem';
 import { privateKeyToAccount } from 'viem/accounts';
 import { anvilFirstPrivateKey } from '../../shared/chains/Anvil';
-import { IVincentNetworkContext } from '../shared/IVincentNetworkContext';
-import { vincentDiamondAddress, vincentSignatures } from './vincent-signatures';
 import { ChrnoicleYellowstone } from '../../shared/chains/ChrnoicleYellowstone';
+import { IVincentNetworkContext } from '../shared/VincentChainClient/IVincentNetworkContext';
+import { vincentDiamondAddress, vincentSignatures } from './vincent-signatures';
 
 export const vincentMainnetNetworkContext: IVincentNetworkContext<
   typeof vincentSignatures
 > = {
   network: 'datil',
   rpcUrl: ChrnoicleYellowstone.rpcUrls[0],
-  privateKey: anvilFirstPrivateKey,
   chainConfig: {
     chain: ChrnoicleYellowstone.chainConfig(),
     contractData: vincentSignatures,
@@ -22,6 +21,3 @@ export const vincentMainnetNetworkContext: IVincentNetworkContext<
     account: privateKeyToAccount(anvilFirstPrivateKey),
   }),
 };
-
-export type VincentDatilMainnetNetworkContext =
-  typeof vincentMainnetNetworkContext;
