+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/README.md
@@ -1,0 +1,1 @@
+a shared directory for all datil networks
