+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/VincentChainClient/apis/highLevelApis/README.md
@@ -1,0 +1,1 @@
+supposed to contain all higher level apis, abstraction over the primitive smart contract functions.
