+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/VincentChainClient/apis/rawContractApis/VincentAppFacet/index.ts
@@ -1,0 +1,8 @@
+// Export all implemented functions
+export * from './registerApp';
+export * from './registerNextAppVersion';
+export * from './enableAppVersion';
+export * from './addAuthorizedRedirectUri';
+export * from './removeAuthorizedRedirectUri';
+export * from './addDelegatee';
+export * from './removeDelegatee';
