+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/VincentChainClient/apis/rawContractApis/VincentAppViewFacet/getAppByDelegatee.spec.ts
@@ -1,4 +1,4 @@
-import { vincentNetworkContext } from '../../../_vincentConfig';
+import { vincentNetworkContext } from '../../../vincentNetworkContext';
 import { getTestContext } from '../testContext';
 import { getAppByDelegatee } from './getAppByDelegatee';
