+++ packages/vincent-contracts/src.ts/networks/vDatil/shared/VincentChainClient/apis/rawContractApis/VincentAppViewFacet/getAppByDelegatee.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { VincentNetworkContext } from '../../../_vincentConfig';
+import { VincentNetworkContext } from '../../../vincentNetworkContext';
 import { createVincentContracts } from '../../utils/createVincentContracts';
 import { toEthAddress } from '../../../../../../shared/utils/z-transformers';
